Move duration fmting into new `value/duration.rs`
This is a `git blame` tollbooth! Move along!
